+++ src/beast/core/parameter/BooleanParameter.java
@@ -89,7 +89,7 @@ public class BooleanParameter extends Parameter.Base<java.lang.Boolean> {
      * Loggable implementation follows *
      */
     @Override
-    public void log(int sampleNr, PrintStream out) {
+    public void log(long sampleNr, PrintStream out) {
         BooleanParameter var = (BooleanParameter) getCurrent();
         int valueCount = var.getDimension();
         for (int i = 0; i < valueCount; i++) {
